LayoutTests/editing/editability/attr-invalid-string.html

<html>
<head>
<link rel="stylesheet" href="../../fast/js/resources/js-test-style.css">
<script src="../../fast/js/resources/js-test-pre.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></p>
<div id="result"></div>
<div id="div" contentEditable="abc"></div>
<script>
description('When contentEditable attribute is invalid string, element.contentEditable returns "inherit".')

shouldBe('document.getElementById("div").getAttribute("contentEditable")','"abc"');
shouldBe('document.getElementById("div").contentEditable', '"inherit"');
debug("FIXME: contentEditable attribute fails to take invalid value. https://bugs.webkit.org/show_bug.cgi?id=50636");
shouldBe('document.getElementById("div").isContentEditable', 'false');
shouldBe('window.getComputedStyle(div, "").getPropertyValue("-webkit-user-modify")', '"read-only"');

document.getElementById("div").style.display= 'none';

window.successfullyParsed = true;
</script>
<script src="../../fast/js/resources/js-test-post.js"></script>
</body>
</html>
